Add Trusted App tests

+++ test/e2e/entity/actionTypes.spec.js
@@ -1,8 +1,6 @@
 const { expect } = require('chai');
 
-const TYPE_NAME = `_actionType${Date.now()}`;
-
-module.exports = (scope, isOperator) => {
+module.exports = (scope, scopeType) => {
   before(() => {
     scope = scope();
   });
@@ -17,9 +15,9 @@ module.exports = (scope, isOperator) => {
       expect(res).to.have.length.gte(5);
     });
 
-    if (isOperator) {
+    if (['operator', 'trustedApp'].includes(scopeType)) {
       it('should create an action type', async () => {
-        const payload = { name: TYPE_NAME };
+        const payload = { name: `_actionType${Date.now()}` };
         actionType = await scope.actionType().create(payload);
 
         expect(actionType).to.be.an('object');
@@ -31,7 +29,9 @@ module.exports = (scope, isOperator) => {
         expect(res).to.be.an('object');
         expect(res.name).to.equal(actionType.name);
       });
+    }
 
+    if (scopeType === 'operator') {
       it('should update an action type', async () => {
         const payload = { tags: ['updated'] };
         const res = await scope.actionType(actionType.name).update(payload);
@@ -39,7 +39,9 @@ module.exports = (scope, isOperator) => {
         expect(res).to.be.an('object');
         expect(res.tags).to.deep.equal(payload.tags);
       });
+    }
 
+    if (['operator', 'trustedApp'].includes(scopeType)) {
       it('should delete an actionType', async () => {
         await scope.actionType(actionType.name).delete();
       });
